# Conflicts:
#	CHANGELOG.md
#	composer.json
#	src/base/Element.php
#	src/config/app.php
#	src/controllers/EntriesController.php
#	src/controllers/UsersController.php
#	src/elements/Entry.php
#	src/elements/db/ElementQuery.php
#	src/web/assets/cp/dist/js/Craft.min.js
#	src/web/assets/cp/dist/js/Craft.min.js.map

+++ src/controllers/UsersController.php
@@ -1447,7 +1447,7 @@ class UsersController extends Controller
                 'html' => $this->_renderPhotoTemplate($user),
             ]);
         } catch (Throwable $exception) {
-            if (isset($fileLocation)) {
+            if (isset($fileLocation) && file_exists($fileLocation)) {
                 FileHelper::unlink($fileLocation);
             }
